dev-requirements.txt
pylint==2.5.2
pylint-django==2.0.15

-r requirements.txt